Centralize quota-exceeded message building in core
Move buildQuotaExceededMessage into a shared chat/common module and have
ChatService substitute the friendly, plan-/reset-date-aware message for
any result flagged isQuotaExceeded, regardless of provider (agent host,
Copilot CLI, Claude CLI, ...). The agent-host adapter now only maps the
error type to flags; the message is no longer built per-provider.
