+++ src/vs/workbench/contrib/chat/test/browser/agentSessions/stateToProgressAdapter.test.ts
@@ -10,8 +10,7 @@ import { ensureNoDisposablesAreLeakedInTestSuite } from '../../../../../../base/
 import { MessageKind, ToolCallStatus, ToolCallConfirmationReason, ToolResultContentType, TurnState, ResponsePartKind, type ActiveTurn, type ICompletedToolCall, type ToolCallRunningState, type Turn, type ToolCallResponsePart, ToolCallCancellationReason, type Message } from '../../../../../../platform/agentHost/common/state/sessionState.js';
 import { IChatToolInvocation, IChatToolInvocationSerialized, type IChatMarkdownContent, type IChatProgressMessage, type IChatUsage } from '../../../common/chatService/chatService.js';
 import { isToolResultInputOutputDetails, type IToolResultInputOutputDetails, ToolDataSource, ToolInvocationPresentation } from '../../../common/tools/languageModelToolsService.js';
-import { ChatEntitlement } from '../../../../../services/chat/common/chatEntitlementService.js';
-import { turnsToHistory as rawTurnsToHistory, activeTurnToProgress as rawActiveTurnToProgress, toolCallStateToInvocation as rawToolCallStateToInvocation, finalizeToolInvocation as rawFinalizeToolInvocation, updateRunningToolSpecificData as rawUpdateRunningToolSpecificData, errorInfoToChatErrorDetails, buildQuotaExceededMessage } from '../../../browser/agentSessions/agentHost/stateToProgressAdapter.js';
+import { turnsToHistory as rawTurnsToHistory, activeTurnToProgress as rawActiveTurnToProgress, toolCallStateToInvocation as rawToolCallStateToInvocation, finalizeToolInvocation as rawFinalizeToolInvocation, updateRunningToolSpecificData as rawUpdateRunningToolSpecificData, errorInfoToChatErrorDetails } from '../../../browser/agentSessions/agentHost/stateToProgressAdapter.js';
 
 // ---- Helper factories -------------------------------------------------------
 
@@ -1476,58 +1475,6 @@ suite('stateToProgressAdapter', () => {
 		test('generic error carries only the message', () => {
 			const details = errorInfoToChatErrorDetails({ errorType: 'query', message: 'Something went wrong' });
 			assert.deepStrictEqual(details, { message: 'Something went wrong' });
-		});
-
-		test('quota error with context replaces the raw message with a friendly one', () => {
-			const details = errorInfoToChatErrorDetails(
-				{ errorType: 'quota', message: '402 {"error":{"code":"quota_exceeded"}}' },
-				{ entitlement: ChatEntitlement.Free },
-			);
-			assert.deepStrictEqual(details, {
-				message: 'You\'ve reached your monthly credit limit. Upgrade to Copilot Pro or wait for your credits to reset.',
-				isQuotaExceeded: true,
-			});
-		});
-
-		test('non-quota error ignores the quota context', () => {
-			const details = errorInfoToChatErrorDetails(
-				{ errorType: 'query', message: 'Something went wrong' },
-				{ entitlement: ChatEntitlement.Free },
-			);
-			assert.deepStrictEqual(details, { message: 'Something went wrong' });
-		});
-	});
-
-	suite('buildQuotaExceededMessage', () => {
-
-		test('free plan without reset date', () => {
-			assert.strictEqual(
-				buildQuotaExceededMessage({ entitlement: ChatEntitlement.Free }),
-				'You\'ve reached your monthly credit limit. Upgrade to Copilot Pro or wait for your credits to reset.',
-			);
-		});
-
-		test('managed plan points at the admin', () => {
-			assert.strictEqual(
-				buildQuotaExceededMessage({ entitlement: ChatEntitlement.Business }),
-				'You\'ve reached your credit limit. Contact your organization\'s Copilot admin to continue.',
-			);
-		});
-
-		test('anonymous user is prompted to sign in', () => {
-			assert.strictEqual(
-				buildQuotaExceededMessage({ entitlement: ChatEntitlement.Unknown }),
-				'You\'ve reached your monthly credit limit. Sign in to keep going.',
-			);
-		});
-
-		test('reset date is appended when provided', () => {
-			const message = buildQuotaExceededMessage({
-				entitlement: ChatEntitlement.Free,
-				quotaResetDate: '2026-06-30T17:00:00.000Z',
-				quotaResetDateHasTime: true,
-			});
-			assert.ok(message.startsWith('You\'ve reached your monthly credit limit. Upgrade to Copilot Pro or wait until your credits reset on '));
 		});
 	});
 });
